Switched to Action
This way,
- reduced folder dependency, which makes Jesse happy
- no more working around hiding from UI

Action originally wasn't really designed for persistent model objects,
but we've fixed them enough in orgfolder and pipeline multibranch that
this should be feasible now
